Move grins_*_solver.h/C to *_solver.h/C
This has been annoying me for awhile now.
