+++ src/solver/include/grins/steady_mesh_adaptive_solver.h
@@ -26,7 +26,7 @@
 #define GRINS_STEADY_MESH_ADAPTIVE_SOLVER_H
 
 // GRINS
-#include "grins/grins_solver.h"
+#include "grins/solver.h"
 #include "grins/mesh_adaptive_solver_base.h"
 
 namespace GRINS
